Skip the steps to make backups for not existing files

+++ scripts/link.sh
@@ -26,10 +26,12 @@ backup_path=$XDG_CACHE_HOME/dotfiles/$(date "+%Y%m%d-%H%M%S")
 mkdir -p $backup_path
 
 for file in "${linked_files[@]}"; do
-  mkdir -p $backup_path/$(dirname $file)
-  cp $HOME/$file $backup_path/$file &&
-    echo "SUCCESS 'cp $DOTFILES_ROOT/$file $backup_path/$file'" ||
-    echo "FAILURE 'cp $DOTFILES_ROOT/$file $backup_path/$file'" && continue
+  if [ -f $HOME/$file ]; then
+    mkdir -p $backup_path/$(dirname $file)
+    cp $HOME/$file $backup_path/$file &&
+      echo "SUCCESS 'cp $DOTFILES_ROOT/$file $backup_path/$file'" ||
+      echo "FAILURE 'cp $DOTFILES_ROOT/$file $backup_path/$file'"
+  fi
   mkdir -p $HOME/$(dirname $file)
   ln -snf $DOTFILES_ROOT/$file $HOME/$file &&
     echo "SUCCESS 'ln -snf $DOTFILES_ROOT/$file $HOME/$file'" ||
